feat: navbar component separated
