+++ pages/blogs.tsx
@@ -23,16 +23,16 @@ export default function Blog({
       </Head>
       <Container>
         <div className="flex flex-col items-start justify-center max-w-2xl mx-auto">
-          <h3 className="mb-4 text-2xl font-bold tracking-tight md:text-4xl text-white">
+          <h3 className="text-2xl font-bold tracking-tight md:text-4xl text-white">
             All Posts
           </h3>
-          <div className="relative w-full mb-4">
+          <div className="relative w-full my-8">
             <input
               aria-label="Search blogs"
               type="text"
               onChange={(e) => setSearchValue(e.target.value)}
               placeholder="Search blogs"
-              className="block w-full px-4 py-2 border-gray-900 focus:ring-blue-500 focus:border-blue-500 bg-gray-800 text-gray-100"
+              className="block w-full px-4 py-2 border-gray-900 focus:ring-blue-500 focus:border-blue-500 bg-gray-800 text-gray-100 rounded-lg"
             />
             <svg
               className="absolute w-5 h-5 right-3 top-3 text-gray-300"
